requirements.txt
# Core AI/ML Dependencies
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
langgraph==0.0.20
openai==1.6.1

# Web Automation
playwright==1.40.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
requests==2.31.0

# Data Processing
pandas==2.1.4
pydantic==2.5.2
python-dotenv==1.0.0

# File Handling
PyPDF2==3.0.1
Pillow==10.1.0
pathlib2==2.3.7

# Async Support
asyncio==3.4.3
aiofiles==23.2.1

# Logging and Monitoring
structlog==23.2.0
rich==13.7.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development Tools
black==23.12.0
flake8==6.1.0
mypy==1.8.0
pre-commit==3.6.0

# Configuration
pyyaml==6.0.1
toml==0.10.2

# Utilities
click==8.1.7
tqdm==4.66.1
